Refactored source code structure
